+++ README.md
@@ -8,41 +8,7 @@ the user can point to the configuration file by setting the `AGENT_CONFIG_PATH` 
 environment variables and credentials can be loaded from a `.env` file in the root directory. or an alternate path set by the `AGENT_ENV_PATH` environment variable.
 data can be stored in local files or remote SQL and Vector databases. the local file storage path can be set by the `AGENT_DATA_PATH` environment variable (defaults to `./data/`).
 
-# Example configuration
-
-{see the AppConfig class in the src/config.py file}
-```yaml
-chunk_overlap: 20
-chunk_size: 1024
-default_llm:
-  class_name: langchain.chat_models.ChatOpenAI
-  model_name: gpt-3.5-turbo
-  temperature: 0
-default_vector_store:
-  class_name: chroma
-  collection_name: default
-  persist_directory: C:\Users\Yaron Haviv\PycharmProjects\demo-llm-agent\data\chroma
-embeddings:
-  class_name: huggingface
-  model_name: all-MiniLM-L6-v2
-log_level: DEBUG
-pipeline_args: {}
-verbose: false
-```
-
-Minimal environment file (`.env`):
-```shell
-IS_LOCAL_CONFIG=1
-OPENAI_API_KEY=your-key
-OPENAI_API_BASE=https://api.openai.com
-```
-
-Substitute the `OPENAI_API_KEY` and `OPENAI_API_BASE` with your own key and base address.
-
-
 # Getting it to work
-
-Make sure there `.env` file is set before running the following commands, and if needed to make modifications to the config
 
 ## Deploy the controller
 This command will start the API controller server into a local docker container.
@@ -51,17 +17,12 @@ make controller
 ```
 
 ## Initialize the database:
-In order to initialize the database, first we need to have the application's code:
+The database is Initialized when building the controller.
+In order to erase and start fresh, we can simply use the controller's command line interface.
 
 ```shell
-git clone -b feature/genai https://github.com/mlrun/mlrun
+python -m controller.src.main initdb
 ```
-Then we can run:
-```shell
-python -m mlrun.genai.main initdb
-```
-
-> should be done only once, or when we want to erase the DB and start fresh 
 
 ## To start the application's API:
 
@@ -79,21 +40,21 @@ make ui
 
 To ingest data into the vector database:
 ```shell
-python -m mlrun.genai.main ingest -l web https://milvus.io/docs/overview.md
+python -m controller.src.main ingest -l web https://milvus.io/docs/overview.md
 ```
 
 To ask a question:
 ```shell   
-python -m mlrun.genai.main query "whats a vector" 
+python -m controller.src.main query "What is a vector?" 
 ```
 
 
 Full CLI:
 
 ```shell
-python -m mlrun.genai.main
+python -m controller.src.main
 
-Usage: python -m mlrun.genai.main [OPTIONS] COMMAND [ARGS]...
+Usage: python -m controller.src.main [OPTIONS] COMMAND [ARGS]...
 
 Options:
   --help  Show this message and exit.
