+++ controller/src/api.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
+import json
 from typing import List, Optional, Tuple, Union
 
 import requests
@@ -21,7 +21,7 @@ from fastapi.middleware.cors import CORSMiddleware
 from pydantic import BaseModel
 
 from controller.src.config import config
-from controller.src.model import ChatSession, DocCollection, OutputMode, QueryItem, User
+from controller.src.model import ChatSession, DocCollection, OutputMode, User
 from controller.src.sqlclient import client
 
 app = FastAPI()
@@ -65,10 +65,10 @@ def get_auth_user(
     if x_username:
         return AuthInfo(username=x_username, token=token)
     else:
-        return AuthInfo(username="yhaviv@gmail.com", token=token)
+        return AuthInfo(username="guest@example.com", token=token)
 
 
-def send_to_application(path: str, method: str = "POST", request=None, auth=None, **kwargs):
+def _send_to_application(path: str, method: str = "POST", request=None, auth=None, **kwargs):
     """
     Send a request to the application's API.
 
@@ -113,7 +113,7 @@ def run_pipeline(
     request: Request, name: str, auth=Depends(get_auth_user)
 ):
     """This is the query command"""
-    return send_to_application(
+    return _send_to_application(
         path=f"pipeline/{name}/run",
         method="POST",
         request=request,
@@ -129,10 +129,12 @@ def ingest(
     params = {
         "path": path,
         "from_file": from_file,
-        "metadata": metadata,
         "version": version,
     }
-    return send_to_application(
+    if metadata is not None:
+        params["metadata"] = json.dumps(metadata)
+
+    return _send_to_application(
         path=f"collections/{collection}/{loader}/ingest",
         method="POST",
         params=params,
